Create OnboardingStateMachineProvider
