+++ p2p_wallet/Scenes/CreateOrRestoreWallet/CreateWallet/CreateWalletViewModel.swift
@@ -12,38 +12,16 @@ final class CreateWalletViewModel: BaseViewModel {
 
     @Injected var onboardingService: OnboardingService
 
-    init(initialState: CreateWalletFlowState?) {
-        let tKeyFacade: TKeyFacade = available(.mockedTKeyFacade) ?
-            TKeyMockupFacade() :
-            TKeyJSFacade(
-                wkWebView: GlobalWebView.requestWebView(),
-                config: .init(
-                    metadataEndpoint: OnboardingConfig.shared.metaDataEndpoint,
-                    torusEndpoint: OnboardingConfig.shared.torusEndpoint,
-                    torusNetwork: "testnet",
-                    torusVerifierMapping: [
-                        "google": OnboardingConfig.shared.torusGoogleVerifier,
-                        "apple": OnboardingConfig.shared.torusAppleVerifier,
-                    ]
-                )
-            )
-
-        #if !RELEASE
-            let apiGatewayEndpoint = String.secretConfig("API_GATEWAY_DEV")!
-        #else
-            let apiGatewayEndpoint = String.secretConfig("API_GATEWAY_PROD")!
-        #endif
-
-        let apiGatewayClient: APIGatewayClient = available(.mockedApiGateway) ?
-            APIGatewayClientImplMock() :
-            APIGatewayClientImpl(endpoint: apiGatewayEndpoint)
-
+    init(
+        initialState: CreateWalletFlowState?,
+        provider: OnboardingStateMachineProvider = OnboardingStateMachineProviderImpl()
+    ) {
         onboardingStateMachine = .init(
             initialState: initialState,
             provider: .init(
                 authService: AuthServiceBridge(),
-                apiGatewayClient: apiGatewayClient,
-                tKeyFacade: tKeyFacade
+                apiGatewayClient: provider.createApiGatewayClient(),
+                tKeyFacade: provider.createTKeyFacade()
             )
         )
